GEODE-3539: Consolidate CliUtil and DataCommandUtils, DataCommandsUtil (#992)
* GEODE-3539: Consolidate CliUtil and DataCommandUtils, DataCommandsUtil

* remove DataCommandsUtil
* move method from DataCommandUtils to CliUtil if applicable.
* move moethds from DataCommandUtils to individual command if only used by one command
